+++ holberton.h
@@ -109,7 +109,7 @@ int is_printable(char);
 int append_hexa_code(char, char[], int);
 int is_digit(char);
 
-long int convert_size_number(long int num, int size);
-long int convert_size_unsgnd(unsigned long int num, int size);
+long int convert_size_number(long long num, int size);
+long int convert_size_unsgnd(unsigned long long num, int size);
 
 #endif /* HOL_H */
